feat: create separate ts file for reactflow types
